feat: Add callback function to notifications
This commit introduces a new `callback` function to the notification
system. This function can be used to trigger custom events on a
notification. It receives a config object and the notification text as
parameters. The `callback` function is optional and is added to the
`config` object in `config.lua`. The `notify` function in `util.lua` has
been updated to call the `callback` function if it is provided.